service/decision/requirements.txt
absl-py
protobuf==3.20.2
google-cloud-spanner==1.10.0
grpcio-tools
python-dotenv
google-cloud-aiplatform
dm-acme==0.4.0
dm-reverb==0.7.0
dm-sonnet==2.0.1
jax==0.4.1
jaxlib==0.4.1
tensorflow==2.8.0
tensorflow-probability==0.15.0
trfl==1.2.0
gym==0.26.2
google-cloud-logging